Moved tensor factories into a separate file
